requirements.txt

streamlit
chromadb
pymupdf
sentence-transformers
torch
rank-bm25

# Groq + OpenAI-compatible client
groq
openai>=1.30.0

# Optional: evaluation utilities
pandas
datasets
tqdm
python-dotenv

# Optional (only if you want RAGAS full metrics; safe to omit)
ragas
langchain-openai
